chore: Rename windows_eventlog to windows_event_log for consistency
Rename all occurrences of `windows_eventlog` to `windows_event_log`:
- Feature flags in Cargo.toml
- Source directory and module names
- Internal events module
- Configuration type name
- Protocol name
- Documentation URLs
- Checkpoint filename

This ensures consistent naming with underscores between words throughout
the codebase and passes spellcheck validation.
