+++ src/sources/windows_event_log/tests.rs
